# Faster heap comparison (WIP)