--- a/source/query_heap_clean.h
+++ b/source/query_heap_clean.h
@@ -304,48 +304,39 @@ namespace JASS
 					By doing the add first its possible to reduce the "usual" path through the code to a single comparison.  The JASS v1 "usual" path took three comparisons.
 				*/ \
 				*which.pointer() += _score; \
-				if (*which.pointer() >= top_k_lower_bound) \
+				if (*which.pointer() < top_k_lower_bound) \
+					continue; \
+				if (needed_for_top_k > 0) \
 					{ \
-					if (which >= accumulator_pointers[0])			/* == is the case where we're the current bottom of heap so might need to be promoted */ \
-						{	\
-						/*
-							We end up in the top-k, now to work out why.  As this is a rare occurence, we've got a little bit of time on our hands
-						*/ \
-						if (needed_for_top_k > 0) \
+					if (*which.pointer() - _score < top_k_lower_bound) \
+						{ \
+						accumulator_pointers[--needed_for_top_k] = which; \
+						if (needed_for_top_k == 0) \
 							{ \
-							/*
-								the heap isn't full yet - so change only happens if we're a new addition (i.e. the old value was a 0)
-							*/ \
-							if (*which.pointer() - _score < top_k_lower_bound) \
-								{ \
-								accumulator_pointers[--needed_for_top_k] = which; \
-								if (needed_for_top_k == 0) \
-									{ \
-									top_results.make_heap(); \
-									if (top_k_lower_bound != 1) \
-										break; /* We must be using the Oracle, and we must have filled the top-k and so we can stop processing this query. */ \
-									top_k_lower_bound = *accumulator_pointers[0];		/* set the new bottom of heap value */ \
-									} \
-								} \
-							} \
-						else \
-							{ \
-							*which.pointer() -= _score; \
-							if (which < accumulator_pointers[0]) \
-								{ \
-								*which.pointer() += _score;					/* we weren't in there before but we are now so replace element 0 */ \
-								top_results.push_back(which);				/* we're not in the heap so add this accumulator to the heap */ \
-								} \
-							else \
-								{ \
-								auto at = top_results.find(which);		/* we're already in there so find us and reshuffle the beap. */ \
-								*which.pointer() += _score; \
-								top_results.promote(which, at);				/* we're already in the heap so promote this document */ \
-								} \
-							top_k_lower_bound = *accumulator_pointers[0];			/* set the new bottom of heap value */ \
+							top_results.make_heap(); \
+							if (top_k_lower_bound != 1) \
+								break; \
+							top_k_lower_bound = *accumulator_pointers[0]; \
 							} \
 						} \
+					continue; \
 					} \
+				if (*which.pointer() == top_k_lower_bound) \
+					{ \
+					if (which.pointer() < accumulator_pointers[0].pointer()) \
+						continue; \
+					top_results.push_back(which); \
+					top_k_lower_bound = *accumulator_pointers[0]; \
+					continue; \
+					} \
+				if (*which.pointer() - _score < top_k_lower_bound || (*which.pointer() - _score == top_k_lower_bound && which.pointer() < accumulator_pointers[0].pointer())) \
+					top_results.push_back(which); \
+				else \
+					{ \
+					auto at = top_results.find(which); \
+					top_results.promote(which, at); \
+					} \
+				top_k_lower_bound = *accumulator_pointers[0]; \
 				}
 
 			/*
@@ -409,52 +400,7 @@ namespace JASS
 				#pragma GCC unroll 8
 #endif
 				for (DOCID_TYPE *current = buffer; current < end; current++)
-					{
-					ACCUMULATOR_TYPE _score = impact;
-					accumulator_pointer which = &accumulators[*current];
-
-					*which.pointer() += _score;
-
-					if (*which.pointer() < top_k_lower_bound)
-						continue;
-					if (needed_for_top_k > 0)
-						{
-						if (*which.pointer() - _score < top_k_lower_bound)
-							{
-							accumulator_pointers[--needed_for_top_k] = which;
-							if (needed_for_top_k == 0)
-								{
-								top_results.make_heap();
-								if (top_k_lower_bound != 1)
-									break;
-								top_k_lower_bound = *accumulator_pointers[0];
-								}
-							}
-						continue;
-						}
-					if (*which.pointer() == top_k_lower_bound)
-						{
-						if (which.pointer() < accumulator_pointers[0].pointer())
-							continue;
-						top_results.push_back(which);
-						top_k_lower_bound = *accumulator_pointers[0];
-						continue;
-						}
-					if (*which.pointer() - _score < top_k_lower_bound)
-						{
-						top_results.push_back(which);
-						}
-					else if (*which.pointer() - _score == top_k_lower_bound && which.pointer() < accumulator_pointers[0].pointer())
-						{
-						top_results.push_back(which);
-						}
-					else
-						{
-						auto at = top_results.find(which);
-						top_results.promote(which, at);
-						}
-					top_k_lower_bound = *accumulator_pointers[0];
-					}
+					ADD_RSV(*current, impact);
 				}
 
 			/*