Improves the performance of accumulator comparison for heap entry. Timings are for wall clock time when executing 60,000 queries on Gov2 (25 million documents)

| Before | After |
| - | - |
| 36m57s | 35m51s |

Testing:
* Same ranking as before when run on 500 queries