@andrewtrotman 2023-09-11
Thanks - its a bit clearer now too.